requirements.txt
appnope==0.1.2
backcall==0.2.0
black==22.1.0
click==8.0.4
dataclasses==0.8
decorator==5.1.1
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.3
ipython==7.16.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
parso==0.7.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.28
ptyprocess==0.7.0
Pygments==2.11.2
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.32
toml==0.10.2
tomli==1.2.3
traitlets==4.3.3
typed-ast==1.5.2
typing_extensions==4.1.1
wcwidth==0.2.5
Werkzeug==2.0.3
zipp==3.6.0
